transparent background
Works well with only with Kitty.

iTerm2 has the problem of not clearing what's between the image and the
next cell. Sixel would have the same problem, but transparency does not
really work anyway; the icy_sixel library has some notion of
transparency in the palette (`sixel_dither::set_transparent`) but I
don't know how to get it to work.