Image widget takes a mutable protocol for kitty transmit
Kitty becomes notably laggy if the image is transmitted on each render.

Kitty's protocol is really built around the idea that an image is
transmitted once, and then "placed" when moving around (unicode
placeholders in our case). To be able to do that, `Image::new` must take
a mutable reference, so that it can track wether the "transmit" sequence
has been sent once.