demo: get stateful resized area
This is just a demo of a workaround to get the resized area, so that the
image can be aligned.

* It's extremely verbose, needs access to too many things, and access to
  otherwise unnecessary structs (ImageSource).
* It does the needs_resize calculation twice per render - but at least
  it doesn't actually resize the image twice.

There should either be a way to just set an alignment option to `Image`,
or there should be a method to get the "resized render area" of the
image. This would be more flexible for future use cases, but also more
cumbersome (to calculate alignments yourself).