ThreadProtocol uses a RenderId to discard stale updates
Fixes #79

Fixes the following scenario:

1. The thread is not blocking on incoming request, unlike the "async"
   example.
2. A large image needs to resize so it's sent off to the processing thread.
3. The app replaces the image with a smaller image which is also sent to
   the processing thread.
4. The large image is processed after the new small image.
5. The large image is rendered which is incorrect.

Track with a `RenderId` counter with a generation, which must be passed
back from the thread along with the protocol to the renamed method
`update_protocol` (was `set_protocol`).

When setting a new image with the new method `set_new_protocol`, the
`RenderId` generation is incremented.