remove chafa-libload
It causes a lot of headaches for nothing.
If the library would be available at runtime, then it should also be
there at compile time. I don't think it makes much sense to distribute
some binary that can load chafa at runtime without linking, as that
would most likely just mean that it won't be loaded.

Build demo and screenshot-tests with chafa-dyn.

No chafa for windows.